backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
requests
langchain
langchain-openai
langgraph
python-dotenv
